fix: fixed some errors and refactored a little
Added BadRequest error to set_commands handling since it was ignoring it

Also added a new message to the same command, it is sent at the start of the process so that the user know the bot is doing something

In manage_permission_list added a negative filter to ignore commands  since the /cancel command was being accepted as an index to remove a user. it isn't an index.